% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oalex.R
\name{oalex_authors}
\alias{oalex_authors}
\title{Search OpenAlex for authors}
\usage{
oalex_authors(search, per_page = 10)
}
\arguments{
\item{search}{Search query string (author name)}

\item{per_page}{Number of results. Default 10.}
}
\value{
tibble: id, display_name, works_count, cited_by_count,
institution, orcid
}
\description{
Search OpenAlex for authors
}
