% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oalex.R
\name{oalex_institutions}
\alias{oalex_institutions}
\title{Search OpenAlex for institutions}
\usage{
oalex_institutions(search, per_page = 10)
}
\arguments{
\item{search}{Search query string (institution name)}

\item{per_page}{Number of results. Default 10.}
}
\value{
tibble: id, display_name, type, country_code, works_count,
cited_by_count
}
\description{
Search OpenAlex for institutions
}
